Change SimpleNotification to Notification
